requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
groq==0.14.0
gTTS==2.5.4
aiosqlite==0.20.0
pydub==0.25.1
python-dotenv==1.0.1
httpx==0.28.1
